rename PrivEnums to Enums
